# Conflicts:
#	resources/views/donate-now/partials/regional-pool.blade.php
#	resources/views/donate/regional-pool.blade.php

+++ app/Http/Controllers/Admin/BusinessUnitController.php
@@ -34,7 +34,7 @@ class BusinessUnitController extends Controller
 
         if($request->ajax()) {
 
-            $columns = ["code","name","status","created_at"];
+            $columns = ["code","name","status","effdt","linked_bu_code", "created_at"];
             $business_units = BusinessUnit::orderBy($columns[$request->input("order")[0]['column']],$request->input("order")[0]['dir']);
 
             return Datatables::of($business_units)
@@ -67,6 +67,7 @@ class BusinessUnitController extends Controller
                 'name' => $request->name,
                 'status' => $request->status,
                 'effdt' => $request->effdt,
+                'linked_bu_code' => $request->linked_bu_code,
                 'notes' => $request->notes,
                 'created_by_id' => Auth::id(),
                 'updated_by_id' => Auth::id(),
@@ -138,14 +139,27 @@ class BusinessUnitController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
-        $business_unit = BusinessUnit::where('id', $id)->first();
-        $business_unit->updated_by_id = Auth::Id();
-        $business_unit->save();
-        
-        $business_unit->delete();
 
-        return response()->noContent();
+        if($request->ajax()) {
+            $business_unit = BusinessUnit::where('id', $id)->first();
+
+            if ($business_unit->hasPledge() ) {
+                return response()->json([
+                    'title'  => "Invalid delete!",
+                    'message' => 'The Business Unit "' . $business_unit->code . ' - ' . $business_unit->name . '" cannot be deleted, it is being referenced on the pledge(s).'], 403);
+            }
+
+            $business_unit->updated_by_id = Auth::Id();
+            $business_unit->save();
+            
+            $business_unit->delete();
+
+            return response()->noContent();
+
+        } else {
+            abort(404);
+        }
     }
 }
